Move Kinect handling to a separate class (KinectInterface).
KinectInterface is a singleton class that is created when the plugin wants
to get data from the Kinect.  On first creation, it spawns a new thread to
handle data from the Kinect.  Plugin objects can register to recieve a
pointer to new data via a callback.  When new data is available, the thread
calls a method in KinectInterface that collects the data and gives a pointer
to the new data to each plugin object that requested it.  The plugin object
then raises a javascript event through the JSAPI object, letting the page
scripts know that new data is available.  All pointers to skeleton data are
smart pointers, so deallocation should be automatic.  Likewise, all pointers
to KinectInterface are smart, so when there are no references to it, its
destructor will run and shut down the Kinect.